fix(tests): isolate email-resend rate-limit tests into their own fixture
The rate-limit tests in AdminControllerIntegrationTests deliberately skipped
TestDataHelper.ClearDatabaseAsync to keep target-user ids unique, but 8
sibling tests in the same class still called it. Since ClearDatabaseAsync
does EnsureDeleted+EnsureCreated on the shared in-memory database, it resets
the auto-increment id sequence — letting a freshly created target user reuse
an id a prior test had already exhausted against the process-wide
emailResendLimiter singleton, which never resets. Moving the 4 rate-limit
tests into their own IClassFixture<WebApplicationFactoryBase> class gives
them a dedicated database and DI container (and thus a fresh rate-limiter
instance), removing the collision instead of relying on test execution order.
