+++ QuestBoard.IntegrationTests/Controllers/AdminControllerIntegrationTests.cs
@@ -387,146 +387,6 @@ public class AdminControllerIntegrationTests : IClassFixture<WebApplicationFacto
         return resolvedUserId!.Value;
     }
 
-    private async Task<HttpResponseMessage> PostSendConfirmationEmailAsync(HttpClient adminClient, int targetUserId)
-    {
-        var getResponse = await adminClient.GetAsync("/Admin/Users", TestContext.Current.CancellationToken);
-        var (token, cookieValue) = await AntiForgeryHelper.ExtractAntiForgeryTokenAsync(getResponse);
-        if (!string.IsNullOrEmpty(cookieValue))
-        {
-            adminClient.DefaultRequestHeaders.Remove("Cookie");
-            adminClient.DefaultRequestHeaders.Add("Cookie", $".AspNetCore.Antiforgery={cookieValue}");
-        }
-
-        var formContent = AntiForgeryHelper.CreateFormContentWithAntiForgeryToken(
-            new Dictionary<string, string> { ["userId"] = targetUserId.ToString() },
-            token);
-
-        return await adminClient.PostAsync("/Admin/SendConfirmationEmail", formContent, TestContext.Current.CancellationToken);
-    }
-
-    // The 4th resend to the SAME target user within the 1-hour window
-    // must be rejected with 429. Program.cs's emailResendLimiter is a process-wide singleton
-    // PartitionedRateLimiter<int> keyed "email-resend:{userId}", PermitLimit=3/hour — this test
-    // uses a fresh, unique target userId so no sibling test's exhausted budget bleeds in.
-    [Fact]
-    public async Task SendConfirmationEmail_ExceedingRateLimit_ShouldReturn429()
-    {
-        // Arrange — deliberately NOT calling ClearDatabaseAsync: it does EnsureDeleted+EnsureCreated,
-        // which resets auto-increment IDs, while the emailResendLimiter singleton is process-wide and
-        // NEVER resets. Wiping the DB risks a freshly-created target user reusing an integer id that
-        // a prior test already exhausted the budget for. A unique userId (guaranteed by not resetting
-        // the id sequence) is the correct isolation mechanism, per this plan's isolation note.
-        var (adminClient, _) = await AuthenticationHelper.CreateAuthenticatedAdminClientAsync(_factory);
-
-        var uniqueSuffix = Guid.NewGuid().ToString("N")[..8];
-        var targetUserId = await CreateUnconfirmedTargetUserAsync(
-            $"emailrate01_{uniqueSuffix}@example.com", "Email Rate 01 Target");
-
-        // Act — 4 rapid resends to the same target user.
-        var responses = new List<HttpResponseMessage>();
-        for (var i = 1; i <= 4; i++)
-        {
-            responses.Add(await PostSendConfirmationEmailAsync(adminClient, targetUserId));
-        }
-
-        // Assert — at least one of the 4 rapid requests was rejected with 429 (mirrors the
-        // ForgotPassword rate-limit analog's "at least one" assertion for robustness).
-        responses.Should().Contain(r => r.StatusCode == HttpStatusCode.TooManyRequests);
-    }
-
-    // Two DISTINCT target users must have INDEPENDENT resend budgets —
-    // proving the partition key is the target userId (not e.g. the requesting admin or a
-    // shared/global bucket, which would be the RESEARCH Pitfall 1 "unknown partition" collapse).
-    [Fact]
-    public async Task SendConfirmationEmail_DifferentTargetUsers_ShouldHaveIndependentBudgets()
-    {
-        // Arrange — no ClearDatabaseAsync (see isolation note above); unique target userIds are
-        // guaranteed by the untouched auto-increment sequence, not by wiping the database.
-        var (adminClient, _) = await AuthenticationHelper.CreateAuthenticatedAdminClientAsync(_factory);
-
-        var uniqueSuffix = Guid.NewGuid().ToString("N")[..8];
-        var targetUserIdA = await CreateUnconfirmedTargetUserAsync(
-            $"emailrate02a_{uniqueSuffix}@example.com", "Email Rate 02 Target A");
-        var targetUserIdB = await CreateUnconfirmedTargetUserAsync(
-            $"emailrate02b_{uniqueSuffix}@example.com", "Email Rate 02 Target B");
-
-        // Act — 3 resends for A (within budget), then 3 for B (within budget), then a 4th for A.
-        var responsesA = new List<HttpResponseMessage>();
-        for (var i = 1; i <= 3; i++)
-        {
-            responsesA.Add(await PostSendConfirmationEmailAsync(adminClient, targetUserIdA));
-        }
-
-        var responsesB = new List<HttpResponseMessage>();
-        for (var i = 1; i <= 3; i++)
-        {
-            responsesB.Add(await PostSendConfirmationEmailAsync(adminClient, targetUserIdB));
-        }
-
-        var fourthForA = await PostSendConfirmationEmailAsync(adminClient, targetUserIdA);
-
-        // Assert — each target user's first-3 sequence is 429-free (independent budgets),
-        // while A's 4th request over its own budget is rejected, proving B's separate budget
-        // was never touched by A's requests.
-        responsesA.Should().NotContain(r => r.StatusCode == HttpStatusCode.TooManyRequests,
-            "target user A's first 3 resends should be within its own budget");
-        responsesB.Should().NotContain(r => r.StatusCode == HttpStatusCode.TooManyRequests,
-            "target user B's first 3 resends should be within its own independent budget");
-        fourthForA.StatusCode.Should().Be(HttpStatusCode.TooManyRequests,
-            "target user A's 4th resend should exceed its own budget, proving the partition key is per-target-user");
-    }
-
-    // EditUser's email-change path shares the same per-target-user
-    // rate limit as SendConfirmationEmail (only email-changing saves are counted).
-    [Fact]
-    public async Task EditUser_EmailChange_ExceedingRateLimit_ShouldReturn429()
-    {
-        // Arrange — no ClearDatabaseAsync (see isolation note above); unique target userId is
-        // guaranteed by the untouched auto-increment sequence, not by wiping the database.
-        var (adminClient, _) = await AuthenticationHelper.CreateAuthenticatedAdminClientAsync(_factory);
-
-        var uniqueSuffix = Guid.NewGuid().ToString("N")[..8];
-        var targetUserId = await CreateUnconfirmedTargetUserAsync(
-            $"emailrate03_{uniqueSuffix}@example.com", "Email Rate 03 Target");
-
-        async Task<HttpResponseMessage> PostEditUserWithEmailChangeAsync(string newEmail)
-        {
-            // Note: EditUser's GET action parameter is named "userId" (not "id"), so the
-            // default {controller}/{action}/{id?} route pattern cannot bind it positionally —
-            // it must be passed as a query string parameter.
-            var getResponse = await adminClient.GetAsync($"/Admin/EditUser?userId={targetUserId}", TestContext.Current.CancellationToken);
-            var (token, cookieValue) = await AntiForgeryHelper.ExtractAntiForgeryTokenAsync(getResponse);
-            if (!string.IsNullOrEmpty(cookieValue))
-            {
-                adminClient.DefaultRequestHeaders.Remove("Cookie");
-                adminClient.DefaultRequestHeaders.Add("Cookie", $".AspNetCore.Antiforgery={cookieValue}");
-            }
-
-            var formContent = AntiForgeryHelper.CreateFormContentWithAntiForgeryToken(
-                new Dictionary<string, string>
-                {
-                    ["Id"] = targetUserId.ToString(),
-                    ["Name"] = "Email Rate 03 Target",
-                    ["Email"] = newEmail,
-                    ["HasKey"] = "false"
-                },
-                token);
-
-            return await adminClient.PostAsync("/Admin/EditUser", formContent, TestContext.Current.CancellationToken);
-        }
-
-        // Act — 4 rapid email-changing saves for the same target user, each with a distinct
-        // new email so ModelState stays valid and the emailChanged branch is entered every time.
-        var responses = new List<HttpResponseMessage>();
-        for (var i = 1; i <= 4; i++)
-        {
-            responses.Add(await PostEditUserWithEmailChangeAsync($"emailrate03_changed{i}_{uniqueSuffix}@example.com"));
-        }
-
-        // Assert — at least one of the 4 rapid email-changing saves was rejected with 429.
-        responses.Should().Contain(r => r.StatusCode == HttpStatusCode.TooManyRequests);
-    }
-
     // DeleteUser must remove only the active-group membership row — the account itself
     // and any other group memberships must survive (SAFE-01).
     [Fact]
@@ -627,47 +487,4 @@ public class AdminControllerIntegrationTests : IClassFixture<WebApplicationFacto
         assertContext.PlayerSignups.Any(s => s.Id == signup.Id).Should().BeTrue("signup rows must not be cascade-deleted");
     }
 
-    // CreateUser's one-shot automated welcome email is explicitly EXEMPT
-    // from the resend rate limit — 4 distinct new-account creations must never 429.
-    [Fact]
-    public async Task CreateUser_RapidRequests_ShouldNotBeRateLimited()
-    {
-        // Arrange
-        await TestDataHelper.ClearDatabaseAsync(_factory.Services);
-        var (adminClient, _) = await AuthenticationHelper.CreateAuthenticatedAdminClientAsync(_factory);
-
-        var uniqueSuffix = Guid.NewGuid().ToString("N")[..8];
-
-        async Task<HttpResponseMessage> PostCreateUserAsync(int index)
-        {
-            var getResponse = await adminClient.GetAsync("/Admin/CreateUser", TestContext.Current.CancellationToken);
-            var (token, cookieValue) = await AntiForgeryHelper.ExtractAntiForgeryTokenAsync(getResponse);
-            if (!string.IsNullOrEmpty(cookieValue))
-            {
-                adminClient.DefaultRequestHeaders.Remove("Cookie");
-                adminClient.DefaultRequestHeaders.Add("Cookie", $".AspNetCore.Antiforgery={cookieValue}");
-            }
-
-            var formContent = AntiForgeryHelper.CreateFormContentWithAntiForgeryToken(
-                new Dictionary<string, string>
-                {
-                    ["Email"] = $"emailrate04_{index}_{uniqueSuffix}@example.com",
-                    ["Name"] = $"Email Rate 04 User {index}",
-                    ["GroupRole"] = ((int)GroupRole.Player).ToString()
-                },
-                token);
-
-            return await adminClient.PostAsync("/Admin/CreateUser", formContent, TestContext.Current.CancellationToken);
-        }
-
-        // Act — 4 distinct new-user creations in rapid succession.
-        var responses = new List<HttpResponseMessage>();
-        for (var i = 1; i <= 4; i++)
-        {
-            responses.Add(await PostCreateUserAsync(i));
-        }
-
-        // Assert — none of the 4 distinct CreateUser POSTs were rate-limited (exempted).
-        responses.Should().NotContain(r => r.StatusCode == HttpStatusCode.TooManyRequests);
-    }
 }
